Refactor to use cuda_compat, and not the unhippified version
